course/serializers.py
from course.models import Course
from rest_framework import serializers


class CourseSerializer(serializers.HyperlinkedModelSerializer):
    class Meta:
        model = Course
        fields = ['code', 'name', 'issue_date', 'begin_date', 'end_date']
